requirements.txt
absl-py==0.15.0
astunparse==1.6.3
audioread==3.0.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
flatbuffers==1.12
fonttools==4.55.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.6.3
numba==0.56.4
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.2
pandas==1.4.4
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.10.1
six==1.15.0
soundfile==0.12.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing-extensions==3.7.4.3
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.12.1
